+++ History.md
@@ -1,7 +1,12 @@
+# TODO:
+
+  * fix TODO's in localStorage fallback.
+
 ## 0.5.0 / 2013-07-..
 
   * ~ merge with weak-indexed
   * ~ use testling to test browsers support
+  * ~ update Readme
 
 ### 0.4.1 / 2013-06-02